r2_seil_common/lambda/gam-to-sf/delivery-reporting/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "GAM to SF Sync Lambda Service"
authors = ["Gerardo Cisneros Mendoza <gerardo.cisneros@contino.io>"]

[tool.poetry.dependencies]
python = "~3.9"
aws-xray-sdk = "^2.9.0"
requests = "^2.27.1"
# setuptools = "^60.8.1"
dependency-injector = "^4.38.0"
gam-core = { path = "../../../lib/gam", develop = true }
salesforce-core = { path = "../../../lib/salesforce", develop = true }
attrs = "^22.1.0"
cryptography = "^38.0.0"

[tool.poetry.dev-dependencies]
pytest = "=7.1.3"
boto3 = "<1.26.100"
pytest-cov = "^3.0.0"
pytest-mock = "^3.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
